refactor(server): Move ObjectTreeReader to zally-core #560
