+++ server/src/test/java/de/zalando/zally/rule/NullPointerExceptionTest.kt
@@ -5,6 +5,7 @@ import com.fasterxml.jackson.databind.JsonNode
 import com.fasterxml.jackson.databind.node.ArrayNode
 import com.fasterxml.jackson.databind.node.ObjectNode
 import de.zalando.zally.core.EMPTY_JSON_POINTER
+import de.zalando.zally.core.ObjectTreeReader
 import de.zalando.zally.core.plus
 import de.zalando.zally.core.toEscapedJsonPointer
 import io.swagger.util.Yaml
