+++ server/src/test/java/de/zalando/zally/rule/zally/AvoidXZallyIgnoreRuleTest.kt
@@ -1,7 +1,7 @@
 package de.zalando.zally.rule.zally
 
-import de.zalando.zally.rule.ObjectTreeReader
 import de.zalando.zally.test.ZallyAssertions
+import de.zalando.zally.core.ObjectTreeReader
 import org.intellij.lang.annotations.Language
 import org.junit.Test
